Extract shared getAllowedOrigins helper for CORS and CSRF.
Keeps origin allowlists from drifting between middleware modules.
